+++ sdk/resourcemanager/computelimit/armcomputelimit/guestsubscriptions_client.go
@@ -42,7 +42,7 @@ func NewGuestSubscriptionsClient(subscriptionID string, credential azcore.TokenC
 // Create - Adds a subscription as a guest to consume the compute limits shared by the host subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-08-15
+// Generated from API version 2026-03-20
 //   - location - The name of the Azure region.
 //   - guestSubscriptionID - The name of the GuestSubscription
 //   - resource - Resource create parameters.
@@ -90,7 +90,7 @@ func (client *GuestSubscriptionsClient) createCreateRequest(ctx context.Context,
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-08-15")
+	reqQP.Set("api-version", "2026-03-20")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	req.Raw().Header["Content-Type"] = []string{"application/json"}
@@ -112,7 +112,7 @@ func (client *GuestSubscriptionsClient) createHandleResponse(resp *http.Response
 // Delete - Deletes a subscription as a guest to stop consuming the compute limits shared by the host subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-08-15
+// Generated from API version 2026-03-20
 //   - location - The name of the Azure region.
 //   - guestSubscriptionID - The name of the GuestSubscription
 //   - options - GuestSubscriptionsClientDeleteOptions contains the optional parameters for the GuestSubscriptionsClient.Delete
@@ -158,7 +158,7 @@ func (client *GuestSubscriptionsClient) deleteCreateRequest(ctx context.Context,
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-08-15")
+	reqQP.Set("api-version", "2026-03-20")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -166,7 +166,7 @@ func (client *GuestSubscriptionsClient) deleteCreateRequest(ctx context.Context,
 // Get - Gets the properties of a guest subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2025-08-15
+// Generated from API version 2026-03-20
 //   - location - The name of the Azure region.
 //   - guestSubscriptionID - The name of the GuestSubscription
 //   - options - GuestSubscriptionsClientGetOptions contains the optional parameters for the GuestSubscriptionsClient.Get method.
@@ -212,7 +212,7 @@ func (client *GuestSubscriptionsClient) getCreateRequest(ctx context.Context, lo
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-08-15")
+	reqQP.Set("api-version", "2026-03-20")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -229,7 +229,7 @@ func (client *GuestSubscriptionsClient) getHandleResponse(resp *http.Response) (
 
 // NewListBySubscriptionLocationResourcePager - Lists all guest subscriptions in a location.
 //
-// Generated from API version 2025-08-15
+// Generated from API version 2026-03-20
 //   - location - The name of the Azure region.
 //   - options - GuestSubscriptionsClientListBySubscriptionLocationResourceOptions contains the optional parameters for the GuestSubscriptionsClient.NewListBySubscriptionLocationResourcePager
 //     method.
@@ -272,7 +272,7 @@ func (client *GuestSubscriptionsClient) listBySubscriptionLocationResourceCreate
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2025-08-15")
+	reqQP.Set("api-version", "2026-03-20")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil